+++ hutool-core/src/test/java/com/xiaoleilu/hutool/captcha/LineCaptchaTest.java
@@ -13,7 +13,12 @@ public class LineCaptchaTest {
 	@Test
 	@Ignore
 	public void lineCaptchaTest() {
+		
+		//定义图形验证码的长和宽
 		LineCaptcha lineCaptcha = CaptchaUtil.createLineCaptcha(200, 100);
+		//图形验证码写出，可以写出到文件，也可以写出到流
 		lineCaptcha.write("d:/aaa.png");
+		//验证图形验证码的有效性，返回boolean值
+		lineCaptcha.verify("1234");
 	}
 }