+++ hutool-crypto/src/test/java/com/xiaoleilu/hutool/crypto/test/SymmetricTest.java
@@ -58,7 +58,7 @@ public class SymmetricTest {
 		//解密
 		byte[] decrypt = aes.decrypt(encrypt);
 		
-		Assert.assertEquals(content, StrUtil.str(decrypt, CharsetUtil.CHARSET_UTF_8));
+		Assert.assertEquals(content, StrUtil.utf8Str(decrypt));
 		
 		//加密为16进制表示
 		String encryptHex = aes.encryptHex(content);
@@ -78,7 +78,7 @@ public class SymmetricTest {
 		byte[] encrypt = des.encrypt(content);
 		byte[] decrypt = des.decrypt(encrypt);
 		
-		Assert.assertEquals(content, StrUtil.str(decrypt, CharsetUtil.CHARSET_UTF_8));
+		Assert.assertEquals(content, StrUtil.utf8Str(decrypt));
 		
 		String encryptHex = des.encryptHex(content);
 		String decryptStr = des.decryptStr(encryptHex);
@@ -96,7 +96,25 @@ public class SymmetricTest {
 		byte[] encrypt = des.encrypt(content);
 		byte[] decrypt = des.decrypt(encrypt);
 		
-		Assert.assertEquals(content, StrUtil.str(decrypt, CharsetUtil.CHARSET_UTF_8));
+		Assert.assertEquals(content, StrUtil.utf8Str(decrypt));
+		
+		String encryptHex = des.encryptHex(content);
+		String decryptStr = des.decryptStr(encryptHex);
+		
+		Assert.assertEquals(content, decryptStr);
+	}
+	
+	@Test
+	public void desdeTest(){
+		String content = "test中文";
+		
+		byte[] key = SecureUtil.generateKey(SymmetricAlgorithm.DESede.getValue()).getEncoded();
+		
+		SymmetricCrypto des = new SymmetricCrypto(SymmetricAlgorithm.DESede, key);
+		byte[] encrypt = des.encrypt(content);
+		byte[] decrypt = des.decrypt(encrypt);
+		
+		Assert.assertEquals(content, StrUtil.utf8Str(decrypt));
 		
 		String encryptHex = des.encryptHex(content);
 		String decryptStr = des.decryptStr(encryptHex);