+++ hutool-http/src/main/java/com/xiaoleilu/hutool/http/HttpConnection.java
@@ -383,6 +383,7 @@ public class HttpConnection {
 	 */
 	public HttpConnection setCookie(String cookie) {
 		if (cookie != null) {
+			log.debug("Cookie: {}", cookie);
 			header(Header.COOKIE, cookie, true);
 		}
 		return this;